requirements.txt
## 📄 `requirements.txt`

```txt
pytesseract
Pillow
customtkinter
pyperclip
